Use macros to define keys
Now there are no more magic numbers, but the number of layers and
functions are temporarily reduced.  They can be expanded again by
moving the modifiers up to the high four bits of int keypress.

Currently LAYER() does not work while fn is pressed as releasing fn
resets the layer to the base layer once fn is released.  This may have
been true in the magic number version as well.